third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-svgdom-markerWidth-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'markerWidth' property of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.markerWidth.baseVal.value is 0
PASS markerElement.markerWidth.baseVal.value is 2
PASS successfullyParsed is true

TEST COMPLETE
